refactor: simplify into seperate packages for each data structure
