--- non_linear/map_set.go
@@ -1,15 +1,0 @@
-package non_linear
-
-type Set map[interface{}]bool
-
-func (s *Set) Add(key interface{}) {
-	(*s)[key] = true
-}
-
-func (s *Set) Remove(key interface{}) {
-	delete((*s), key)
-}
-
-func (s *Set) Find(key interface{}) bool {
-	return (*s)[key]
-}
