--- non_linear/map_set_test.go
@@ -1,16 +1,0 @@
-package non_linear
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestMapSet(t *testing.T) {
-	set := make(Set)
-	set.Add(1)
-	set.Add(2)
-	fmt.Println(set.Find(1))
-	set.Remove(1)
-	fmt.Println(set.Find(1))
-
-}
